refactor: prepare for v2
